InventoryManagement/requirements.txt
asgiref==3.8.1
Django==4.0.10
django-model-observer==1.0.0
django-postgres-partitioning==1.1.0
psycopg2-binary==2.9.10
six==1.16.0
sqlparse==0.5.2
django-import-export==3.0.0
